[indexer][editor] Move editor tests from indexer to Editor project
