+++ editor/editor_tests/CMakeLists.txt
@@ -8,6 +8,8 @@ set(
   feature_matcher_test.cpp
   match_by_geometry_test.cpp
   opening_hours_ui_test.cpp
+  osm_editor_test.cpp
+  osm_editor_test.hpp
   ui2oh_test.cpp
   user_stats_test.cpp
   xml_feature_test.cpp
@@ -19,6 +21,19 @@ omim_link_libraries(
   ${PROJECT_NAME}
   editor
   platform_tests_support
+  generator_tests_support
+  generator
+  indexer_tests_support
+  indexer
+  routing
+  routing_common
+  search
+  succinct
+  storage
+  tess2
+  jansson
+  icu
+  protobuf
   platform
   geometry
   coding
