+++ editor/editor_tests/osm_editor_test.hpp
